requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-phone-field==1.8.1
django-webpack-loader==1.6.0
djangorestframework==3.14.0
google-auth==2.14.1
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
whitenoise==6.2.0
